requirements.txt
# Python dependencies for PDF processing
PyMuPDF>=1.23.0
pdf-redactor>=0.1.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
Pillow>=10.0.0
pytesseract>=0.3.10

